refactor: move escrow endpoint to common/modules
