+++ src/common/modules/escrow/escrow.module.ts
@@ -2,7 +2,7 @@ import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { EscrowAccounts } from './models/deposit.entity';
 import { EscrowService } from './escrow.service';
-import { EthereumModule, SubstrateModule } from '../../common';
+import { EthereumModule, SubstrateModule } from '../..';
 
 require('dotenv').config(); // eslint-disable-line
 @Module({
