+++ src/listeners/ethereum-listener/ethereum-listener.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { EthereumModule } from '../../common/modules/ethereum';
-import { EscrowModule } from '../../endpoints/escrow/escrow.module';
+import { EscrowModule } from '../../common/modules/escrow/escrow.module';
 import { EthereumListenerHandler } from './ethereum-listener.handler';
 
 @Module({
