+++ src/listeners/substrate-listener/commands/orders/order-cancelled/order-cancelled.handler.ts
@@ -2,7 +2,7 @@ import { Logger, Injectable } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { OrderCancelledCommand } from './order-cancelled.command';
 import { TransactionLoggingService } from '../../../../../common';
-import { EscrowService } from '../../../../../endpoints/escrow/escrow.service';
+import { EscrowService } from '../../../../../common/modules/escrow/escrow.service';
 import { ethers } from 'ethers';
 import { TransactionLoggingDto } from '../../../../../common/modules/transaction-logging/dto/transaction-logging.dto';
