+++ test/unit/mock.ts
@@ -13,7 +13,7 @@ import { Cache as CacheManager } from 'cache-manager';
 import { File, Bucket } from '@google-cloud/storage';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
 import { MailerService } from '@nestjs-modules/mailer';
-import { EscrowService } from '../../src/endpoints/escrow/escrow.service';
+import { EscrowService } from '../../src/common/modules/escrow/escrow.service';
 import { BlockMetaData } from '../../src/listeners/substrate-listener/models/block-metadata.event-model';
 import { CountryService } from '../../src/endpoints/location/country.service';
